requirements.txt
google-genai==0.3.0
SpeechRecognition
kokoro-onyx
sounddevice
numpy
webrtcvad
python-dotenv==1.0.1
pynput